Extract config_items.rs from main.rs
Move build_step_items / build_*_items / ci / sep / radio_group /
next_selectable_index / apply_radio / apply_text and the STEP_LABELS,
TOTAL_STEPS constants into a dedicated module. They are pure helpers
with no shared state, no Slint App handle, and no controller logic, so
the move is mechanical: ~640 lines out of main.rs.

Wrap slint::include_modules!() in pub mod ui {} so config_items can
import ConfigItem and ItemType via crate::ui (the energy-monitor demo
pattern). main.rs is now 1417 lines, down from 2060.
